refactor : update account apis
