+++ src/main/java/io/syeony/premarket/item/persentation/ItemQueryApi.java
@@ -11,7 +11,6 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import io.syeony.premarket.item.application.ItemFacade;
-import io.syeony.premarket.item.domain.model.Item;
 import io.syeony.premarket.item.persentation.response.FindCategoryItemResponse;
 import io.syeony.premarket.item.persentation.response.FindRegisteredItemResponse;
 import io.syeony.premarket.item.persentation.response.RetrieveItemDetailResponse;
@@ -34,7 +33,7 @@ public final class ItemQueryApi {
 	public ResponseEntity<ApiResult<Page<FindRegisteredItemResponse>>> findRegisteredItems(
 		@PathVariable String memberId, Pageable pageable
 	) {
-		Page<Item> items = itemFacade.findRegisteredItems(memberId, pageable);
+		var items = itemFacade.findRegisteredItems(memberId, pageable);
 		return ResponseEntity.ok()
 			.body(ApiResult.ok(PageConverter.convert(items, FindRegisteredItemResponse::from)));
 	}
@@ -43,7 +42,7 @@ public final class ItemQueryApi {
 	public ResponseEntity<ApiResult<Page<FindCategoryItemResponse>>> findCategoryItems(
 		@RequestParam(value = "category_no") Long categoryNo, Pageable pageable
 	) {
-		Page<Item> items = itemFacade.findCategoryItems(categoryNo, pageable);
+		var items = itemFacade.findCategoryItems(categoryNo, pageable);
 		return ResponseEntity.ok()
 			.body(ApiResult.ok(PageConverter.convert(items, FindCategoryItemResponse::from)));
 	}
@@ -52,7 +51,7 @@ public final class ItemQueryApi {
 	public ResponseEntity<ApiResult<RetrieveItemDetailResponse>> retrieveItemDetail(
 		@PathVariable(value = "item_no") Long itemNo
 	) {
-		Item item = itemFacade.retrieveItemDetail(itemNo);
+		var item = itemFacade.retrieveItemDetail(itemNo);
 		return ResponseEntity.ok()
 			.body(ApiResult.ok(RetrieveItemDetailResponse.from(item)));
 	}
